db_sync/pyproject.toml
[tool.poetry]
name = "db_sync"
version = "0.1.0"
description = ""
authors = ["JusbrasilDevs"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiokafka = "0.11.0"
asyncpg = "0.29.0"
yoyo-migrations = "^8.2.0"
psycopg2-binary = "2.9.9"
setuptools = ">=80.9.0,<81.0.0"
pydantic = "^2.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0,<9.0"
pytest-asyncio = ">=0.23,<0.24"
